fix(test): make 9P lcreate mode portable to macOS
libc::S_IFREG is u16 on macOS but u32 on Linux; the typed flush test only
compiled on Linux.

+++ zerofs/src/rpc/server/typed_flush_tests.rs
@@ -494,7 +494,8 @@ async fn unix_ninep_standard_and_verified_fsync_wait_for_real_remote_writeback()
             fid,
             b"typed-durability",
             (libc::O_RDWR | libc::O_CREAT) as u32,
-            libc::S_IFREG | 0o644,
+            // u32::from: S_IFREG is u16 on macOS, u32 on Linux.
+            u32::from(libc::S_IFREG) | 0o644,
             0,
         )
         .await
